feat(app): separate app db management from cache management
refactor(data): split out orchestrators from data sources
